backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-multipart==0.0.12
pdfminer.six==20231228
python-dotenv==1.0.1
PyGithub==2.4.0
requests==2.32.3